switch file id to 32bit and get sram push working

+++ avr/usbload/dir.c
@@ -34,12 +34,12 @@ void dir_entry_start(){
 }
 
 void dir_entry_dump(uint32_t addr, dir_ent_t* ent){
-    debug(DEBUG_FAT,"dir_entry_dump: addr=0x%06lx id=%i name=%s size=%li attr=%i\n", addr, ent->id, ent->file_name, 
+    debug(DEBUG_FAT,"dir_entry_dump: addr=0x%06lx id=%li name=%s size=%li attr=%i\n", addr, ent->id, ent->file_name, 
                     ent->file_size, ent->file_attr);
 }
 
 
-void dir_entry_add(uint16_t id, uint8_t* file_name,uint32_t file_size,uint8_t file_attr){
+void dir_entry_add(uint32_t id, uint8_t* file_name,uint32_t file_size,uint8_t file_attr){
     uint32_t addr;
     dir_ent_t ent;
     strncpy(ent.file_name,file_name,13);
@@ -57,5 +57,6 @@ void dir_entry_header(uint16_t position, uint8_t * header){
     addr = DIR_ENTRY_LOC + ( position << DIR_ENTRY_SIZE_SHIFT ) + DIR_ENTRY_HEADER_OFF;
     sram_bulk_copy(addr, (uint8_t *) header, DIR_ENTRY_HEADER_SIZE);
 }
+
